Drive rotation with the keyboard
This commit adds the Keyboard package, which we use to turn into
Rotation messages.

This also tidies up some SVG messes, though I think there's gotta be
a better way to rotate the board around the origin. The rotate property
accepts an x and y point to rotate around, but we need to calculate the
origin manually in order to do so. We also have to offset the group
manually.