Animate board rotation
This uses the wonderful elm-style-animation library to rotate the board.
It seems to work a lot like React Motion: you provide a spring to define
how an interpolation function will animate a select number of CSS
properties. You then send a message, passing it a desired final state,
and the Animation.update function will figure out how to animate the
object to the desired state.

Something that seems a bit awkward right now is that the Turnabout
module is the one that knows all about how to rotate the board, while it
really seems like that should be the responsibility of the view. Perhaps
View is the wrong name for that module... maybe it should be Board?
Board could then export its own model, view, and update function that
Turnabout could then compose into the full app.