Add buttons
This adds two button to rotate the board, one to undo. I want to get
this on a touch-device and see if the animation strategy is too taxing
on a phone.

This also makes dealing with color a little more type-friendly. Weird
that there's no color conversion library in the standard lib though.