Compress javascript in build
This commit adds some bulid dependencies that rely on NPM.
`script/publish` and `script/server` act as they did before, but they
employ NPM to provide a version of elm-make and elm-reactor, and
compress the final javascript using the closure compiler.

One nice side effect of this is that if NPM is installed on the machine
(seems highly likely in this world...), then either of those script
commands will Just Work.