Move animation state into Board.elm
This is a refactoring that moves the state keeping track of the
animation of the board into the Board module itself. This way, animation
info doesn't leak into Turnabout.elm.

This tries to follow the methodology seen in [elm-sortable-table], but
I need to study the code in that library a little more fully. This is
also the path taken in [elm-tutorial], but it seems like
elm-sortable-table has the nicer api.

The main benefit of this approach is that you never need to think about
*how* Board will animate to a new one, only that is exposes some state
that we mount top-level and interact with using a Board.update function.

I think there's likely a better way of _triggering_ the animation than
by issuing a command, but I haven't figured it out yet. It will involve
a closer look into the style-element library.

[elm-sortable-table]: https://github.com/evancz/elm-sortable-table
[elm-tutorial]: https://www.elm-tutorial.org/en-v01/02-elm-arch/06-composing.html