Rudimentary level select
Now you can select a level from a homescreen.

Not sure yet whether I should be stuffing the level in the model, or
computing it like I am now by storing the number of the currently
selected item. I think it can always be computed state... I can always
generate the current state of the board from the level start and the
list of rotations. That may be too expensive? Time will tell.