Add test harness
This sets up elm-test... it's a bit of a bear to work with so far.
Really not a fan of the testing DSL. The rspec style doesn't feel like
it's a very good match for Elm's syntax.

It's enough to setup testing for a little helper function I'm working on
to build a version of the Level data that takes into account parts that
can move and parts that are just pieces of the board. Ideal solution
likely involves a better parser that can create the entire Level record
as it chews through a level string.