Update Playfield without an additional message
I read a very good article explaining why my "send" command was bad:

https://medium.com/elm-shorts/how-to-turn-a-msg-into-a-cmd-msg-in-elm-5dd095175d84

Formerly, I was taking another trip through the Elm runtime in order to
call it's update function. The rotate, select level, and end level
actions were creating a command, and then piggybacking on the
PlayfieldMessage message to call the update function. Now the update
function is called directly, and the state is set in one fell swoop.