Restrict movement until the end of animation
This adds a check in the main subscriptions setup --- if the board is
still animating, then it removes the keyboard subscription. Once the
board settles, then it allows keyboard events again.

I originally fought the compiler for a little while trying to peek into
the value of isRunning coming from the elm-style-animation library. Each
animation model keeps track of whether it's running or not, and it
seemed like a good way to accomplish this feature. I couldn't unwrap the
value though, I don't think the library exports the type contructor for
an Animation value. In the end, I opted for a bit of new Playfield state
that's just a boolean keeping track of whether the animation is going or
not.

This solution might be better though for a few reasons:

- it gives us more control into what counts as a running animation. It
doesn't really matter that there's an animation for scaling values, for
example.

- it will be easier and cheaper to monitor whether multiple animations
are going on (think of the rotation + update situation).