Create Board module
Previously Level was the owner of the data structures for all kinds of
tiles. Over time though, it seemed that the board was something a bit
different from the rest of the tiles... because it's immutable during
the course of a level, and it's mostly queried for available positions,
it makes sense as a stand-alone data structure.

One change that rippled throughout the codebase a bit was moving the
size of the board from Level into Board. I think this is the right place
for it --- the size is also fixed for the course of the level, and
should never go out-of-sync, so why not associate the two alongside
operations to the board?

Creating a data structure like this allows us to create query and update
methods apart from the Level record. This is useful in parsing, and will
be especially useful in the level editor.