Move level data-related functions into Level.Model
Having Level.Types was always a little awkward. It existed so that
Level.Parser and Level could share definitions. Some bad parts of that
setup: `empty` was defined in Parser, Level held a lot of information
about how parsing worked.

Something not great is how much the tests and code reference internals
of level. I think the solution there is a new abstraction around
representing positions and the current direction of gravity.