Hacky (visual) blocks implementation
I'm taking advantage of the fact that many blocks are rectangular here
by rendering an svg square that's the width and height of the _extent_
that a block occupies. Something particularly nice though is the zipDict
function, which is like finding the union of a set, but for dictionary
values. This lets us more easily make a list of blocks and their
positions, so that we have something more easily drawable