Draw marbles based on animated position properties
There's nothing yet to drive the animation, but this commit adds another
layer of marbles to the board, where their placement comes from the dict
of animated positions!

I'm not a huge fan of the separation I've created so far. It's really
awkward to send the Playfield model these messges, and it seems like it
has a lot of potential to go out-of-sync.