Animate marble movement
This is some very rudimentary animation for moving the marble, but it's
better than nothing.

Unfortunately I think the idea of the spring-based animation breaks down
in this case. There are three interpolation values, but none of them are
the right thing to use in this case. Spring is going to pull at the same
rate, so two marbles will land at the same time, one smoothly from
a small distance, one strongly from a larger one. Plus the idea is all
wrong: a spring decelerates as it falls, whereas gravity would
accelerate. Easings offer some hope, but the problem with easings is
that it's on *us* to figure out how long it should take. Possible! But
irritating. Speed would be great if we could define a constant for
acceleration.

I think maybe implementing some form of animation myself might be nice.
There's nothing that prevents me from pre-computing the positions every
second along the way, and running a bunch of keyframed animations. Or
keep a dictionary of animatable properties keyed by id, and compute
distance traveled myself each step --- knowing dt from
requestAnimationFrame makes that do-able.