Level.positionOf can fail
There's an important invariant to maintain when performing operations
around looking up block locations, and that's that a positionable
element (marble, goal, block) must have a an associated position at all
times. So long as an instance of a marble is around and can be rendered,
it needs to have an associated position in the lookup table. This commit
maintains that invariant by turning positionOf into a Result. Should it
fail, we should handle that case. Right now, the app just crashes if
that state ever happens.