Create Marble module
The main reason for doing this is that the Movable representation of
marble doesn't make much sense. It has the coordinate as a part of its
data! Because this value changes over time, it makes more sense to
represent only the necessary data that needs to be associated with
a marble: its color.

The pattern established with the blocks is a good example of this.
Blocks only have identity, and it's the Level.positions mapping that
associates the identity (an Id in this case) with the coordinate.

Another benefit of turning Marble into a module is that we can group the
type alongside the view for the thing. This also means that we can treat
all of the code that positions/animates the marble as a separate `svg`
element with width and height properties.