Add routing for levels
This commit adds routing! For now, the routes are:

- /           --- LevelSelect
- /level/<id> --- Level id

This lets us treat transitioning between pages in the game as
transitions between URLs. The Navigation package uses pushState to
intercept link transitions. This lets us treat the level we're on as
a function of the URL, which is one less thing to keep track of in the
model.