Store page state top-level
This adds a new type, Page, that exists to hold the state for each major
route in the app. Right now, that's only Playfield. LevelSelect doesn't
have any information, and Error exists for the case when you request
a level that doesn't exist.

Next step will be moving the keyboard subscription, moves, and rotation
stuff into Playfield.