Put controls into Playfield
This moves a lot of the code around handling keypresses, turning them into
a list of moves, and applying animations to the Playfield module. This
makes sense, since we're not doing anything with keypresses, moves, etc
in Turnabout.elm. Sometimes we have a level, sometimes we don't at that
layer, but with this approach, there's always a level in Playfield. This
was annoying too because we kept having to send the Playfield module
messages whenever the state changed.